rag_system/requirements.txt
chromadb>=0.4.0
sentence-transformers
pypdf
langchain
langchain-community
groq
python-dotenv
tqdm
transformers
torch
accelerate
